+++ build.gradle
@@ -11,6 +11,7 @@ buildscript {
         classpath 'com.google.gms:google-services:4.3.5'
         classpath 'com.google.firebase:perf-plugin:1.3.5'
         classpath 'com.google.firebase:firebase-crashlytics-gradle:2.5.1'
+        classpath 'androidx.navigation:navigation-safe-args-gradle-plugin:2.3.4'
     }
 }
