+++ firestore/app/src/main/java/com/google/firebase/example/fireeats/java/MainFragment.java
@@ -1,0 +1,341 @@
+package com.google.firebase.example.fireeats.java;
+
+import android.app.Activity;
+import android.content.DialogInterface;
+import android.content.Intent;
+import android.os.Bundle;
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.StringRes;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.text.HtmlCompat;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.navigation.NavAction;
+import androidx.navigation.NavController;
+import androidx.navigation.NavOptions;
+import androidx.navigation.NavOptionsBuilder;
+import androidx.navigation.Navigation;
+import androidx.navigation.fragment.NavHostFragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+
+import com.firebase.ui.auth.AuthUI;
+import com.firebase.ui.auth.ErrorCodes;
+import com.firebase.ui.auth.IdpResponse;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.android.material.snackbar.Snackbar;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.example.fireeats.R;
+import com.google.firebase.example.fireeats.databinding.FragmentMainBinding;
+import com.google.firebase.example.fireeats.java.adapter.RestaurantAdapter;
+import com.google.firebase.example.fireeats.java.model.Rating;
+import com.google.firebase.example.fireeats.java.model.Restaurant;
+import com.google.firebase.example.fireeats.java.util.RatingUtil;
+import com.google.firebase.example.fireeats.java.util.RestaurantUtil;
+import com.google.firebase.example.fireeats.java.viewmodel.MainActivityViewModel;
+import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.FirebaseFirestoreException;
+import com.google.firebase.firestore.Query;
+import com.google.firebase.firestore.WriteBatch;
+
+import java.util.Collections;
+import java.util.List;
+
+public class MainFragment extends Fragment implements
+        FilterDialogFragment.FilterListener,
+        RestaurantAdapter.OnRestaurantSelectedListener, View.OnClickListener {
+
+    private static final String TAG = "MainActivity";
+
+    private static final int RC_SIGN_IN = 9001;
+
+    private static final int LIMIT = 50;
+
+    private FragmentMainBinding mBinding;
+
+    private FirebaseFirestore mFirestore;
+    private Query mQuery;
+
+    private FilterDialogFragment mFilterDialog;
+    private RestaurantAdapter mAdapter;
+
+    private MainActivityViewModel mViewModel;
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        mBinding = FragmentMainBinding.inflate(inflater, container, false);
+        return mBinding.getRoot();
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+        mBinding.filterBar.setOnClickListener(this);
+        mBinding.buttonClearFilter.setOnClickListener(this);
+
+        // View model
+        mViewModel = new ViewModelProvider(this).get(MainActivityViewModel.class);
+
+        // Enable Firestore logging
+        FirebaseFirestore.setLoggingEnabled(true);
+
+        // Firestore
+        mFirestore = FirebaseFirestore.getInstance();
+
+        // Get ${LIMIT} restaurants
+        mQuery = mFirestore.collection("restaurants")
+                .orderBy("avgRating", Query.Direction.DESCENDING)
+                .limit(LIMIT);
+
+        // RecyclerView
+        mAdapter = new RestaurantAdapter(mQuery, this) {
+            @Override
+            protected void onDataChanged() {
+                // Show/hide content if the query returns empty.
+                if (getItemCount() == 0) {
+                    mBinding.recyclerRestaurants.setVisibility(View.GONE);
+                    mBinding.viewEmpty.setVisibility(View.VISIBLE);
+                } else {
+                    mBinding.recyclerRestaurants.setVisibility(View.VISIBLE);
+                    mBinding.viewEmpty.setVisibility(View.GONE);
+                }
+            }
+
+            @Override
+            protected void onError(FirebaseFirestoreException e) {
+                // Show a snackbar on errors
+                Snackbar.make(mBinding.getRoot(),
+                        "Error: check logs for info.", Snackbar.LENGTH_LONG).show();
+            }
+        };
+
+        mBinding.recyclerRestaurants.setLayoutManager(new LinearLayoutManager(requireContext()));
+        mBinding.recyclerRestaurants.setAdapter(mAdapter);
+
+        // Filter Dialog
+        mFilterDialog = new FilterDialogFragment();
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+
+        // Start sign in if necessary
+        if (shouldStartSignIn()) {
+            startSignIn();
+            return;
+        }
+
+        // Apply filters
+        onFilter(mViewModel.getFilters());
+
+        // Start listening for Firestore updates
+        if (mAdapter != null) {
+            mAdapter.startListening();
+        }
+    }
+
+    @Override
+    public void onStop() {
+        super.onStop();
+        if (mAdapter != null) {
+            mAdapter.stopListening();
+        }
+    }
+
+    @Override
+    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
+        inflater.inflate(R.menu.menu_main, menu);
+        super.onCreateOptionsMenu(menu, inflater);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case R.id.menu_add_items:
+                onAddItemsClicked();
+                break;
+            case R.id.menu_sign_out:
+                AuthUI.getInstance().signOut(requireContext());
+                startSignIn();
+                break;
+        }
+        return super.onOptionsItemSelected(item);
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+        if (requestCode == RC_SIGN_IN) {
+            IdpResponse response = IdpResponse.fromResultIntent(data);
+            mViewModel.setIsSigningIn(false);
+
+            if (resultCode != Activity.RESULT_OK) {
+                if (response == null) {
+                    // User pressed the back button.
+                    requireActivity().finish();
+                } else if (response.getError() != null
+                        && response.getError().getErrorCode() == ErrorCodes.NO_NETWORK) {
+                    showSignInErrorDialog(R.string.message_no_network);
+                } else {
+                    showSignInErrorDialog(R.string.message_unknown);
+                }
+            }
+        }
+    }
+
+    public void onFilterClicked() {
+        // Show the dialog containing filter options
+        mFilterDialog.show(getChildFragmentManager(), FilterDialogFragment.TAG);
+    }
+
+    public void onClearFilterClicked() {
+        mFilterDialog.resetFilters();
+
+        onFilter(Filters.getDefault());
+    }
+
+    @Override
+    public void onRestaurantSelected(DocumentSnapshot restaurant) {
+        // Go to the details page for the selected restaurant
+        MainFragmentDirections.ActionMainFragmentToRestaurantDetailFragment action = MainFragmentDirections
+                .actionMainFragmentToRestaurantDetailFragment(restaurant.getId());
+
+        NavHostFragment.findNavController(this)
+                .navigate(action);
+    }
+
+    @Override
+    public void onFilter(Filters filters) {
+        // Construct query basic query
+        Query query = mFirestore.collection("restaurants");
+
+        // Category (equality filter)
+        if (filters.hasCategory()) {
+            query = query.whereEqualTo(Restaurant.FIELD_CATEGORY, filters.getCategory());
+        }
+
+        // City (equality filter)
+        if (filters.hasCity()) {
+            query = query.whereEqualTo(Restaurant.FIELD_CITY, filters.getCity());
+        }
+
+        // Price (equality filter)
+        if (filters.hasPrice()) {
+            query = query.whereEqualTo(Restaurant.FIELD_PRICE, filters.getPrice());
+        }
+
+        // Sort by (orderBy with direction)
+        if (filters.hasSortBy()) {
+            query = query.orderBy(filters.getSortBy(), filters.getSortDirection());
+        }
+
+        // Limit items
+        query = query.limit(LIMIT);
+
+        // Update the query
+        mAdapter.setQuery(query);
+
+        // Set header
+        mBinding.textCurrentSearch.setText(HtmlCompat.fromHtml(filters.getSearchDescription(requireContext()),
+                HtmlCompat.FROM_HTML_MODE_LEGACY));
+        mBinding.textCurrentSortBy.setText(filters.getOrderDescription(requireContext()));
+
+        // Save filters
+        mViewModel.setFilters(filters);
+    }
+
+    private boolean shouldStartSignIn() {
+        return (!mViewModel.getIsSigningIn() && FirebaseAuth.getInstance().getCurrentUser() == null);
+    }
+
+    private void startSignIn() {
+        // Sign in with FirebaseUI
+        Intent intent = AuthUI.getInstance().createSignInIntentBuilder()
+                .setAvailableProviders(Collections.singletonList(
+                        new AuthUI.IdpConfig.EmailBuilder().build()))
+                .setIsSmartLockEnabled(false)
+                .build();
+
+        startActivityForResult(intent, RC_SIGN_IN);
+        mViewModel.setIsSigningIn(true);
+    }
+
+    private void onAddItemsClicked() {
+        // Add a bunch of random restaurants
+        WriteBatch batch = mFirestore.batch();
+        for (int i = 0; i < 10; i++) {
+            DocumentReference restRef = mFirestore.collection("restaurants").document();
+
+            // Create random restaurant / ratings
+            Restaurant randomRestaurant = RestaurantUtil.getRandom(requireContext());
+            List<Rating> randomRatings = RatingUtil.getRandomList(randomRestaurant.getNumRatings());
+            randomRestaurant.setAvgRating(RatingUtil.getAverageRating(randomRatings));
+
+            // Add restaurant
+            batch.set(restRef, randomRestaurant);
+
+            // Add ratings to subcollection
+            for (Rating rating : randomRatings) {
+                batch.set(restRef.collection("ratings").document(), rating);
+            }
+        }
+
+        batch.commit().addOnCompleteListener(new OnCompleteListener<Void>() {
+            @Override
+            public void onComplete(@NonNull Task<Void> task) {
+                if (task.isSuccessful()) {
+                    Log.d(TAG, "Write batch succeeded.");
+                } else {
+                    Log.w(TAG, "write batch failed.", task.getException());
+                }
+            }
+        });
+    }
+
+    private void showSignInErrorDialog(@StringRes int message) {
+        AlertDialog dialog = new AlertDialog.Builder(requireContext())
+                .setTitle(R.string.title_sign_in_error)
+                .setMessage(message)
+                .setCancelable(false)
+                .setPositiveButton(R.string.option_retry, new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialogInterface, int i) {
+                      startSignIn();
+                    }
+                })
+                .setNegativeButton(R.string.option_exit, new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialogInterface, int i) {
+                        requireActivity().finish();
+                    }
+                }).create();
+
+        dialog.show();
+    }
+
+    @Override
+    public void onClick(View v) {
+        switch (v.getId()) {
+            case R.id.filterBar:
+                onFilterClicked();
+                break;
+            case R.id.buttonClearFilter:
+                onClearFilterClicked();
+                break;
+        }
+    }
+}
