+++ firestore/app/src/main/java/com/google/firebase/example/fireeats/kotlin/MainFragment.kt
@@ -1,0 +1,277 @@
+package com.google.firebase.example.fireeats.kotlin
+
+import android.app.Activity
+import android.content.Intent
+import android.os.Bundle
+import android.util.Log
+import android.view.*
+import androidx.annotation.StringRes
+import androidx.appcompat.app.AlertDialog
+import androidx.core.text.HtmlCompat
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.fragment.findNavController
+import androidx.recyclerview.widget.LinearLayoutManager
+import com.firebase.ui.auth.AuthUI
+import com.firebase.ui.auth.ErrorCodes
+import com.firebase.ui.auth.IdpResponse
+import com.google.android.material.snackbar.Snackbar
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.example.fireeats.R
+import com.google.firebase.example.fireeats.databinding.FragmentMainBinding
+import com.google.firebase.example.fireeats.kotlin.adapter.RestaurantAdapter
+import com.google.firebase.example.fireeats.kotlin.model.Restaurant
+import com.google.firebase.example.fireeats.kotlin.util.RatingUtil
+import com.google.firebase.example.fireeats.kotlin.util.RestaurantUtil
+import com.google.firebase.example.fireeats.kotlin.viewmodel.MainActivityViewModel
+import com.google.firebase.firestore.DocumentSnapshot
+import com.google.firebase.firestore.FirebaseFirestore
+import com.google.firebase.firestore.FirebaseFirestoreException
+import com.google.firebase.firestore.Query
+import com.google.firebase.firestore.ktx.firestore
+import com.google.firebase.ktx.Firebase
+
+class MainFragment : Fragment(),
+        FilterDialogFragment.FilterListener,
+        RestaurantAdapter.OnRestaurantSelectedListener {
+
+    lateinit var firestore: FirebaseFirestore
+    lateinit var query: Query
+
+    private lateinit var binding: FragmentMainBinding
+    private lateinit var filterDialog: FilterDialogFragment
+    lateinit var adapter: RestaurantAdapter
+
+    private lateinit var viewModel: MainActivityViewModel
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        binding = FragmentMainBinding.inflate(inflater, container, false);
+        return binding.root;
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        // View model
+        viewModel = ViewModelProvider(this).get(MainActivityViewModel::class.java)
+
+        // Enable Firestore logging
+        FirebaseFirestore.setLoggingEnabled(true)
+
+        // Firestore
+        firestore = Firebase.firestore
+
+        // Get ${LIMIT} restaurants
+        query = firestore.collection("restaurants")
+                .orderBy("avgRating", Query.Direction.DESCENDING)
+                .limit(LIMIT.toLong())
+
+        // RecyclerView
+        adapter = object : RestaurantAdapter(query, this@MainFragment) {
+            override fun onDataChanged() {
+                // Show/hide content if the query returns empty.
+                if (itemCount == 0) {
+                    binding.recyclerRestaurants.visibility = View.GONE
+                    binding.viewEmpty.visibility = View.VISIBLE
+                } else {
+                    binding.recyclerRestaurants.visibility = View.VISIBLE
+                    binding.viewEmpty.visibility = View.GONE
+                }
+            }
+
+            override fun onError(e: FirebaseFirestoreException) {
+                // Show a snackbar on errors
+                Snackbar.make(binding.root,
+                        "Error: check logs for info.", Snackbar.LENGTH_LONG).show()
+            }
+        }
+
+        binding.recyclerRestaurants.layoutManager = LinearLayoutManager(context)
+        binding.recyclerRestaurants.adapter = adapter
+
+        // Filter Dialog
+        filterDialog = FilterDialogFragment()
+
+        binding.filterBar.setOnClickListener { onFilterClicked() }
+        binding.buttonClearFilter.setOnClickListener { onClearFilterClicked() }
+    }
+
+    public override fun onStart() {
+        super.onStart()
+
+        // Start sign in if necessary
+        if (shouldStartSignIn()) {
+            startSignIn()
+            return
+        }
+
+        // Apply filters
+        onFilter(viewModel.filters)
+
+        // Start listening for Firestore updates
+        adapter.startListening()
+    }
+
+    public override fun onStop() {
+        super.onStop()
+        adapter.stopListening()
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
+        inflater.inflate(R.menu.menu_main, menu)
+        return super.onCreateOptionsMenu(menu, inflater)
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when (item.itemId) {
+            R.id.menu_add_items -> onAddItemsClicked()
+            R.id.menu_sign_out -> {
+                AuthUI.getInstance().signOut(requireContext())
+                startSignIn()
+            }
+        }
+        return super.onOptionsItemSelected(item)
+    }
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        if (requestCode == RC_SIGN_IN) {
+            val response = IdpResponse.fromResultIntent(data)
+            viewModel.isSigningIn = false
+
+            if (resultCode != Activity.RESULT_OK) {
+                if (response == null) {
+                    // User pressed the back button.
+                    requireActivity().finish()
+                } else if (response.error != null && response.error!!.errorCode == ErrorCodes.NO_NETWORK) {
+                    showSignInErrorDialog(R.string.message_no_network)
+                } else {
+                    showSignInErrorDialog(R.string.message_unknown)
+                }
+            }
+        }
+    }
+
+    private fun onFilterClicked() {
+        // Show the dialog containing filter options
+        filterDialog.show(childFragmentManager, FilterDialogFragment.TAG)
+    }
+
+    private fun onClearFilterClicked() {
+        filterDialog.resetFilters()
+
+        onFilter(Filters.default)
+    }
+
+    override fun onRestaurantSelected(restaurant: DocumentSnapshot) {
+        // Go to the details page for the selected restaurant
+        val action = MainFragmentDirections
+            .actionMainFragmentToRestaurantDetailFragment(restaurant.id)
+
+        findNavController().navigate(action)
+    }
+
+    override fun onFilter(filters: Filters) {
+        // Construct query basic query
+        var query: Query = firestore.collection("restaurants")
+
+        // Category (equality filter)
+        if (filters.hasCategory()) {
+            query = query.whereEqualTo(Restaurant.FIELD_CATEGORY, filters.category)
+        }
+
+        // City (equality filter)
+        if (filters.hasCity()) {
+            query = query.whereEqualTo(Restaurant.FIELD_CITY, filters.city)
+        }
+
+        // Price (equality filter)
+        if (filters.hasPrice()) {
+            query = query.whereEqualTo(Restaurant.FIELD_PRICE, filters.price)
+        }
+
+        // Sort by (orderBy with direction)
+        if (filters.hasSortBy()) {
+            query = query.orderBy(filters.sortBy.toString(), filters.sortDirection)
+        }
+
+        // Limit items
+        query = query.limit(LIMIT.toLong())
+
+        // Update the query
+        adapter.setQuery(query)
+
+        // Set header
+        binding.textCurrentSearch.text = HtmlCompat.fromHtml(filters.getSearchDescription(requireContext()),
+                HtmlCompat.FROM_HTML_MODE_LEGACY)
+        binding.textCurrentSortBy.text = filters.getOrderDescription(requireContext())
+
+        // Save filters
+        viewModel.filters = filters
+    }
+
+    private fun shouldStartSignIn(): Boolean {
+        return !viewModel.isSigningIn && Firebase.auth.currentUser == null
+    }
+
+    private fun startSignIn() {
+        // Sign in with FirebaseUI
+        val intent = AuthUI.getInstance().createSignInIntentBuilder()
+                .setAvailableProviders(listOf(AuthUI.IdpConfig.EmailBuilder().build()))
+                .setIsSmartLockEnabled(false)
+                .build()
+
+        startActivityForResult(intent, RC_SIGN_IN)
+        viewModel.isSigningIn = true
+    }
+
+    private fun onAddItemsClicked() {
+        // Add a bunch of random restaurants
+        val batch = firestore.batch()
+        for (i in 0..9) {
+            val restRef = firestore.collection("restaurants").document()
+
+            // Create random restaurant / ratings
+            val randomRestaurant = RestaurantUtil.getRandom(requireContext())
+            val randomRatings = RatingUtil.getRandomList(randomRestaurant.numRatings)
+            randomRestaurant.avgRating = RatingUtil.getAverageRating(randomRatings)
+
+            // Add restaurant
+            batch.set(restRef, randomRestaurant)
+
+            // Add ratings to subcollection
+            for (rating in randomRatings) {
+                batch.set(restRef.collection("ratings").document(), rating)
+            }
+        }
+
+        batch.commit().addOnCompleteListener { task ->
+            if (task.isSuccessful) {
+                Log.d(TAG, "Write batch succeeded.")
+            } else {
+                Log.w(TAG, "write batch failed.", task.exception)
+            }
+        }
+    }
+
+    private fun showSignInErrorDialog(@StringRes message: Int) {
+        val dialog = AlertDialog.Builder(requireContext())
+                .setTitle(R.string.title_sign_in_error)
+                .setMessage(message)
+                .setCancelable(false)
+                .setPositiveButton(R.string.option_retry) { _, _ -> startSignIn() }
+                .setNegativeButton(R.string.option_exit) { _, _ -> requireActivity().finish() }.create()
+
+        dialog.show()
+    }
+
+    companion object {
+
+        private const val TAG = "MainActivity"
+
+        private const val RC_SIGN_IN = 9001
+
+        private const val LIMIT = 50
+    }
+}
