+++ firestore/build.gradle
@@ -2,14 +2,15 @@
 
 buildscript {
     repositories {
-        jcenter()
         google()
+        jcenter()
         mavenLocal()
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:4.1.2'
         classpath 'com.google.gms:google-services:4.3.5'
         classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.4.31'
+        classpath 'androidx.navigation:navigation-safe-args-gradle-plugin:2.3.4'
     }
 }
